+++ client/src/components/Today/Row.tsx
@@ -1,0 +1,23 @@
+import { ActivityWithIds } from "../../types/server/activity.types";
+import { ID } from "../../types/server/utility.types";
+import Activity from "./Activity";
+import HourMark from "./HourMark";
+import * as S from "./Today.style";
+
+type RowProps = {
+	index: number;
+	activities: ActivityWithIds[];
+	indentation: Map<ID, number>;
+};
+
+export default function Row({ index, activities, indentation }: RowProps) {
+	return (
+		<S.Row>
+			<HourMark index={index} />
+
+			{activities.map((a) => (
+				<Activity key={a.activity_id} activity={a} indentation={indentation} />
+			))}
+		</S.Row>
+	);
+}
