+++ client/src/components/Today/Tasks.tsx
@@ -1,0 +1,23 @@
+import useTagsQuery from "../../lib/use-tags-query";
+import { ActivityWithIds } from "../../types/server/activity.types";
+import Task from "./Task";
+import * as S from "./Today.style";
+
+type TasksProps = {
+	activities: ActivityWithIds[];
+};
+
+export default function Tasks({ activities }: TasksProps) {
+	const { data: tags } = useTagsQuery();
+
+	return (
+		<S.TasksWrapper>
+			<S.BlockTitle>Tasks</S.BlockTitle>
+			<S.Tasks>
+				{activities.map((a) => (
+					<Task key={a.activity_id} activity={a} tagsById={tags?.tagsById} />
+				))}
+			</S.Tasks>
+		</S.TasksWrapper>
+	);
+}
